-Added mixpanel events on user, groups, conversation and welcome views -Added control on Groups page to redirect if you're not signed in and to check if you're part of the group

+++ app/controllers/groups_controller.rb
@@ -1,11 +1,17 @@
 class GroupsController < ApplicationController
   def show
-  	@group = Group.find(params[:id])
-  	@users = @group.users.all
+  	deny_access
 
-  	@posts = @group.posts.order("created_at DESC").all
-    @newpost = @group.posts.new
+  	if Groupuser.group_user_exist?(params[:id], current_user.id)
+	  	@group = Group.find(params[:id])
+	  	@users = @group.users.all
 
-    current_user.notifications.where("notifiable_type = 'Group'").destroy_all
+	  	@posts = @group.posts.order("created_at DESC").all
+	    @newpost = @group.posts.new
+
+	    current_user.notifications.where("notifiable_type = 'Group'").destroy_all
+	else
+		redirect_to "/users"
+	end
   end
 end
